+++ gdtoolkit/formatter/annotation.py
@@ -14,6 +14,7 @@ STANDALONE_ANNOTATIONS = [
     "export_subgroup",
     "icon",
     "tool",
+    "warning_ignore",
 ]